refactor helpers
